Disable flush integration test
Flush is too harsh a method (could have unexpected side effect when the
tests are run, since all data in bucket is deleted).

Also, it can only be performed if the flush feature is enabled on the
bucket's configuration.

So for now this test is disabled (with an Ignore message).

Closes #1