Rework manager construction and allow dynamic creation
This change reworks the constructor of `CouchbaseCacheManager` so that
it is easier to create in the 80% case, without having to provide a Map
of caches.

It introduces CacheBuilder as a way to describe construction of a cache
and perform it lazily.

The CacheManager can be created in three modes:

1. The CacheManager is created with only a cache builder: it is therefore
dynamic and any cache created on the fly uses the specified builder
2. The CacheManager is created with a cache builder and a list of cache
names: it is static and the caches are created on startup using the
specified builder
3. The CacheManager is created with a mapping between a cache name and a
cache builder. It is static as well and each individual cache can have
dedicated settings, including different bucket

It is possible to enable or disable the dynamic mode by specifying a
default CacheBuilder once the CacheManager has been created: specifying
null is allowed and disable the dynamic mode.