Rework CacheManager construction and allow dynamic creation
This change reworks the constructor of `CouchbaseCacheManager` so that
it is easier to create in the 80% case, without having to provide a Map.

One can now create a manager with a statically declared list of caches
in 3 ways:

 * just provide the names and the `Bucket` to use (expected 80% case)
 * provide names, bucket and a common non-zero ttl for all caches
 * provide a map of names and custom distinct ttls + the bucket for case
   where you need static declaration but distinct expirations.

Furthermore, dynamic creation of caches is now allowed. After having
constructed the `CouchbaseCacheManager` without a list of cache names,
one can request any cache and it will be created with the currently set
`defaultTtl` (which can be changed between cache requests).

Closes #4