Polish
This commit updates the current proposal to only use a CacheBuilder
rather than a CacheBuilder and a CacheTemplate. The main difference is
that the name of the cache is no longer an argument of the builder which
allows it to be shared for multiple cache instances.

The CacheManager can be created in three modes:

1. The CacheManager is created with only a cache builder: it is therefore
dynamic and any cache created on the fly uses the specified builder
2. The CacheManager is created with a cache builder and a list of cache
names: it is static and the caches are created on startup using the
specified builder
3. The CacheManager is created with a mapping between a cache name and a
cache builder. It is static as well and each individual cache can have
dedicated settings, including different bucket

It is possible to enable or disable the dynamic mode by specifying a
default CacheBuilder once the CacheManager has been created: specifying
null is allowed and disable the dynamic mode.